requirements.txt
pylatexgit
jsonschema
PyYAML
requests